+++ validator/src/bin/solana-test-validator.rs
@@ -281,6 +281,8 @@ fn main() {
             .map(|v| v.into_iter().collect())
             .unwrap_or_default();
 
+    let clone_feature_set = matches.is_present("clone_feature_set");
+
     let warp_slot = if matches.is_present("warp_slot") {
         Some(match matches.value_of("warp_slot") {
             Some(_) => value_t_or_exit!(matches, "warp_slot", Slot),
@@ -507,6 +509,17 @@ fn main() {
                 .expect("bug: --url argument missing?"),
         ) {
             println!("Error: clone_upgradeable_programs failed: {e}");
+            exit(1);
+        }
+    }
+
+    if clone_feature_set {
+        if let Err(e) = genesis.clone_feature_set(
+            cluster_rpc_client
+                .as_ref()
+                .expect("bug: --url argument missing?"),
+        ) {
+            println!("Error: clone_feature_set failed: {e}");
             exit(1);
         }
     }